+++ Projects/NUCLEO-WB09KE/Examples/RADIO/RADIO_Sniffer/Src/main.c
@@ -32,7 +32,7 @@ typedef struct circular_fifo_s
   uint16_t tail_index;
   uint8_t curr_size;
   uint8_t max_size;
-  uint8_t elem_size;
+  uint16_t elem_size;
   uint8_t *buffer;
 } circular_fifo_t;
 
@@ -86,7 +86,7 @@ static void MX_GPIO_Init(void);
 static void MX_RADIO_Init(void);
 static void MX_RADIO_TIMER_Init(void);
 /* USER CODE BEGIN PFP */
-void fifo_init(circular_fifo_t *fifo, uint8_t *buffer, uint8_t size, uint8_t elem_size);
+void fifo_init(circular_fifo_t *fifo, uint8_t *buffer, uint8_t size, uint16_t elem_size);
 uint8_t fifo_size(circular_fifo_t *fifo);
 uint8_t fifo_put(circular_fifo_t *fifo, uint8_t elem_size, uint8_t elem_data_offset, uint8_t *elem);
 uint8_t fifo_get(circular_fifo_t *fifo, uint8_t *elem);
@@ -121,7 +121,7 @@ int main(void)
   /* Configure the system clock */
   SystemClock_Config();
 
-/* Configure the peripherals common clocks */
+  /* Configure the peripherals common clocks */
   PeriphCommonClock_Config();
 
   /* USER CODE BEGIN SysInit */
@@ -361,7 +361,7 @@ void HAL_RADIO_CallbackRcvError(RxStats_t* rxDataStats)
    /* USER CODE END HAL_RADIO_CallbackRcvError 1 */
 }
 
-void fifo_init(circular_fifo_t *fifo, uint8_t *buffer, uint8_t size, uint8_t elem_size)
+void fifo_init(circular_fifo_t *fifo, uint8_t *buffer, uint8_t size, uint16_t elem_size)
 {
   fifo->tail_index = fifo->head_index = 0;
   fifo->curr_size = 0;
@@ -381,7 +381,7 @@ uint8_t fifo_put(circular_fifo_t *fifo, uint8_t elem_size, uint8_t elem_data_off
   {
     //printf("PUT: Curr Size [%d], Head Index [%d], Tail Index [%d] \n", fifo->curr_size, fifo->head_index, fifo->tail_index);
     /*copy element*/
-    uint8_t fifo_ind = fifo->tail_index * fifo->elem_size;
+    uint16_t fifo_ind = fifo->tail_index * fifo->elem_size;
     for (uint8_t i = 0; i < elem_size; i++)
     {
       fifo->buffer[fifo_ind] = elem[i + elem_data_offset];
@@ -408,7 +408,7 @@ uint8_t fifo_get(circular_fifo_t *fifo, uint8_t *elem)
   //printf("GET: Curr Size [%d], Head Index [%d], Tail Index [%d] \n", fifo->curr_size, fifo->head_index, fifo->tail_index);
   if (fifo->curr_size > 0)
   {
-    uint8_t fifo_ind = fifo->head_index * fifo->elem_size;
+    uint16_t fifo_ind = fifo->head_index * fifo->elem_size;
     for (uint8_t i = 0; i < fifo->elem_size; i++)
     {
       elem[i] = fifo->buffer[fifo_ind];
@@ -435,7 +435,7 @@ uint8_t fifo_get_elem_size(circular_fifo_t *fifo, uint8_t size, uint8_t *elem)
   //printf("GET: Curr Size [%d], Head Index [%d], Tail Index [%d] \n", fifo->curr_size, fifo->head_index, fifo->tail_index);
   if (fifo->curr_size > 0)
   {
-    uint8_t fifo_ind = fifo->head_index * fifo->elem_size;
+    uint16_t fifo_ind = fifo->head_index * fifo->elem_size;
     for (uint8_t i = 0; i < size; i++)
     {
       elem[i] = fifo->buffer[fifo_ind];
